Response handling rework - Response managment has been concolidated into their own namespace instead of being seperate methods spread accross the whole program - Lowered the volatility of the chat response module
