+++ Robit/Response/Auto.cs
@@ -1,0 +1,59 @@
+﻿using DSharpPlus.EventArgs;
+using static Robit.FileManager;
+
+namespace Robit.Response
+{
+    /// <summary>
+    /// Handles autoresponses
+    /// </summary>
+    public static class Auto
+    {
+        /// <summary>
+        /// Checks for saved auto responses and generates an appropriate response
+        /// </summary>
+        /// <param name="messageArgs">Discord message creation arguments</param>
+        /// <returns>
+        /// <list type="table">
+        /// <listheader>A tuple containing response information</listheader>
+        /// <item>
+        /// <list type="table">
+        /// <listheader>Item1 (bool)</listheader>
+        /// <item>True: Generation succeeded</item>
+        /// <item>False: Generation failed</item>
+        /// </list>
+        /// </item>
+        /// <item>
+        /// <list type="table">
+        /// <listheader>Item2 (string)</listheader>
+        /// <item>Generation successful: Generation result</item>
+        /// <item>Generation failiure: Fail reason</item>
+        /// </list>
+        /// </item>
+        /// </list>
+        /// </returns>
+        public static async Task<Tuple<bool, string>> GenerateAutoResponse(MessageCreateEventArgs messageArgs)
+        {
+            List<ResponseManager.ResponseEntry> responseEntries = ResponseManager.ReadEntries(messageArgs.Guild.Id.ToString());
+
+            string messageLower = messageArgs.Message.Content.ToLower();
+
+            messageLower = WordFilter.WordFilter.SpecialCharacterRemoval(messageLower);
+
+            string[] wordsInMessage = messageLower.Split(' ');
+
+            foreach (string word in wordsInMessage)
+            {
+                foreach (ResponseManager.ResponseEntry responseEntry in responseEntries)
+                {
+                    if (word == responseEntry.content.ToLower())
+                    {
+                        await messageArgs.Message.RespondAsync(responseEntry.response);
+                        return Tuple.Create(true, responseEntry.response);
+                    }
+                }
+            }
+
+            return Tuple.Create(false, "No saved matches found");
+        }
+    }
+}
